+++ client/projects/client/src/app/layout/sidebar/sidebar.component.ts
@@ -1,0 +1,11 @@
+import { Component } from "@angular/core";
+import { version } from '../../../../../../package.json';
+
+
+@Component({
+    selector: 'app-sidebar',
+    templateUrl: './sidebar.component.html',
+})
+export class SidebarComponent {
+    version: string = version;
+}