+++ client/projects/lib/src/lib/types/index.ts
@@ -1,3 +1,4 @@
 export * from './api'
 export * from './profile'
 export * from './xmpp'
+export * from './ui'