+++ client/projects/lib/src/public-api.ts
@@ -5,3 +5,6 @@
 export * from './lib/types'
 export * from './lib/utils'
 export * from './lib/lib.module';
+
+export * from './mock/conversations'
+export * from './mock/messages'